Modify SigninOrUp process, and remove accessToken from User class
