+++ backend/src/main/java/org/nexters/cultureland/api/user/controller/UserController.java
@@ -1,15 +1,8 @@
 package org.nexters.cultureland.api.user.controller;
 
-import org.nexters.cultureland.api.user.dto.UserDto;
 import org.nexters.cultureland.api.user.model.User;
-import org.nexters.cultureland.api.user.service.SSOService;
 import org.nexters.cultureland.api.user.service.UserService;
-import org.nexters.cultureland.api.user.service.impl.FacebookSSOServiceImpl;
-import org.nexters.cultureland.api.user.service.impl.KakaoSSOServiceImpl;
 import org.nexters.cultureland.common.ResponseMessage;
-import org.nexters.cultureland.common.excepion.BadRequestException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,22 +10,8 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping(value = "/users")
 public class UserController {
-    @Autowired
-    private ApplicationContext appContext;
 
-    private SSOService ssoService;
     private UserService userService;
-
-    @PostMapping
-    public ResponseEntity<ResponseMessage> signInorSignUp(@RequestParam String snsName,
-                                                          @RequestBody UserDto userDto){
-        if(userDto.getAccessToken() == null) {throw new BadRequestException("No AccessToken");}
-        ssoService = this.getSSOService(snsName);
-        String jwtToken = ssoService.signInOrSignUp(userDto.getAccessToken());
-        ResponseMessage resp = ResponseMessage.getOkResponseMessage();
-        resp.setMessage(jwtToken);
-        return new ResponseEntity<>(resp, HttpStatus.OK);
-    }
 
     @GetMapping(value = "/{userId}")
     public ResponseEntity<ResponseMessage> requestUserInfos(@PathVariable Long userId){
@@ -50,18 +29,7 @@ public class UserController {
         return new ResponseEntity<>(resp, HttpStatus.OK);
     }
 
-    public SSOService getSSOService(String snsName) {
-        SSOService ssoService = null;
-        switch (snsName.toUpperCase()){
-            case "KAKAO":
-                ssoService = appContext.getBean(KakaoSSOServiceImpl.class);
-                break;
-            case "FACEBOOK":
-                ssoService = appContext.getBean(FacebookSSOServiceImpl.class);
-                break;
-        }
-        return ssoService;
-    }
+
 
     public UserController(UserService userService) {
         this.userService = userService;
