+++ backend/src/main/java/org/nexters/cultureland/config/WebConfig.java
@@ -11,7 +11,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(tokenInterceptor())
                 .addPathPatterns("/**")
-                .excludePathPatterns("/users", "/error");
+                .excludePathPatterns("/signInOrUp", "/error");
     }
 
     @Bean
